rag-backend/pyproject.toml
[tool.poetry]
name = "rag-benchmark"
version = "0.1.0"
description = "RAG Benchmarking Platform"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.25.0"
pydantic = "^2.5.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.9"
python-multipart = "^0.0.6"
celery = "^5.3.6"
redis = "^5.0.1"
networkx = "^3.2.1"
langchain = "^0.1.0"
langchain-openai = "^0.0.5"
openai = "^1.6.0"
pgvector = "^0.2.0"
pandas = "^2.1.4"
beautifulsoup4 = "^4.12.2"
pdfplumber = "^0.10.3"
jsonschema = "^4.20.0"
pytest = "^7.4.3"
httpx = "^0.26.0"
prometheus-client = "^0.19.0"
python-jose = "^3.3.0"
python-dotenv = "^1.0.0"
ragas = "^0.1.0"
tiktoken = "^0.5.2"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.8.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 